device_impl: raise msg ID from uint8_t to uint16_t
New mavlink messages come with msg ids > 255. Therefore, we need to
raise the msg id field form uint8_t to uint16_t.

+++ src/device_impl.cpp
@@ -57,7 +57,7 @@ DeviceImpl::~DeviceImpl()
     }
 }
 
-void DeviceImpl::register_mavlink_message_handler(uint8_t msg_id,
+void DeviceImpl::register_mavlink_message_handler(uint16_t msg_id,
                                                   mavlink_message_handler_t callback,
                                                   const void *cookie)
 {
